% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifiedyh.R
\name{systematic_labels}
\alias{systematic_labels}
\title{Systematic Sampling and Labeling}
\usage{
systematic_labels(df, group_col, sampling_interval)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{group_col}{A character string specifying the column to use for grouping.}

\item{sampling_interval}{A numeric value representing the interval for systematic sampling.}
}
\value{
A data frame with an additional column "Systematic_Yes_No" containing the systematically sampled "Yes"/"No" labels.
}
\description{
This function performs systematic sampling on the dataframe and assigns "Yes" or "No" labels to rows based on the specified interval.
}
\examples{
result <- systematic_labels(iris, group_col = "Species", sampling_interval = 2)
}
